+++ libs/iso-bmff/src/readers/readStss.ts
@@ -3,11 +3,11 @@ import type { SyncSampleBox } from '../boxes/SyncSampleBox.ts'
 import type { IsoBoxReadView } from '../IsoBoxReadView.ts'
 
 /**
- * Parse a SyncSampleBox from an IsoView
+ * Parse a `SyncSampleBox` from an `IsoBoxReadView`.
  *
- * @param view - The IsoView to read data from
+ * @param view - The `IsoBoxReadView` to read data from
  *
- * @returns A parsed SyncSampleBox
+ * @returns A parsed `SyncSampleBox`
  *
  * @public
  */