+++ libs/iso-bmff/src/writers/writeDref.ts
@@ -4,12 +4,12 @@ import type { IsoBoxWriteViewConfig } from '../IsoBoxWriteViewConfig.ts'
 import { writeChildBoxes } from '../utils/writeChildBoxes.ts'
 
 /**
- * Write a DataReferenceBox to an IsoDataWriter.
+ * Write a `DataReferenceBox` to an `IsoBoxWriteView`.
  *
- * @param box - The DataReferenceBox fields to write
+ * @param box - The `DataReferenceBox` fields to write
  * @param config - The IsoBoxWriteViewConfig to use
  *
- * @returns An IsoDataWriter containing the encoded box
+ * @returns An `IsoBoxWriteView` containing the encoded box
  *
  * @public
  */